Reduce cogntive complexity in GetProducerBillingInstructionsAsync
